dir changes, list services

+++ scripts/init-common.sh
@@ -6,14 +6,35 @@ if [[ -f "$HOME/.rvm/scripts/rvm" ]]; then
   rvm gemset use ${RVM_GEMSET_NAME_TP_INV}
 fi
 
+function start_kafka {
+			tmux new-window -t TpInv -n kafka "services/kafka.sh start"
+			echo "Waiting for Kafka initialization 20 sec..."
+			sleep 20
+			echo "Done"
+}
+
+function stop_kafka {
+    ./services/kafka.sh stop
+    sleep 1
+    tmux kill-window -t TpInv:kafka &> /dev/null
+}
+
 function start_svc_in_tmux {
     local svc=$1
 
-    tmux new-window -t TpInv -n ${svc} "services/${svc}.sh"
+    if [[ ${svc} == "kafka" ]]; then
+        start_kafka
+    else
+        tmux new-window -t TpInv -n ${svc} "services/${svc}.sh"
+    fi
 }
 
 function stop_svc_in_tmux {
     local svc=$1
 
-    tmux kill-window -t TpInv:${svc} &> /dev/null
+    if [[ ${svc} == "kafka" ]]; then
+        stop_kafka
+    else
+        tmux kill-window -t TpInv:${svc} &> /dev/null
+    fi
 }
